Add absolute file paths
