+++ nodemon.json
@@ -1,0 +1,6 @@
+{
+  "watch": ["src", ".env"],
+  "ext": "js,ts,json",
+  "ignore": ["src/logs/*", "src/**/*.{spec,test}.ts"],
+  "exec": "ts-node -r tsconfig-paths/register --transpile-only src/app.ts"
+}
