+++ package.json
@@ -2,21 +2,22 @@
   "name": "usc-backend",
   "version": "1.0.0",
   "description": "USC backend ptototype",
-  "main": "index.js",
+  "main": "src/app.js",
   "scripts": {
     "test": "jest --coverage",
-    "build": "npx tsc --outDir ./dist",
-    "start": "node dist/index.js",
-    "dev": "concurrently \"npx tsc --watch\" \"nodemon -q dist/index.js\"",
-    "lint": "eslint src/**/*.{js,ts,json}",
-    "lint:fix": "eslint --fix 'src/**/*.{js,ts,json}'",
+    "build": "swc src -d dist --source-maps --copy-files",
+    "start": "npm run build && cross-env NODE_ENV=production node dist/app.js",
+    "dev": "cross-env NODE_ENV=development nodemon",
+    "lint": "eslint --ignore-path .gitignore src/**/*.{js,ts,json}",
+    "lint:fix": "npm run lint -- --fix",
     "format": "prettier --write 'src/**/*.{js,ts,md,json}' --config ./.prettierrc.json",
-    "migrate": "prisma migrate dev",
-    "deploy": "prisma migrate deploy && prisma generate",
-    "seed": "prisma db seed"
+    "prisma:init": "prisma init",
+    "prisma:migrate": "prisma migrate dev --preview-feature",
+    "prisma:seed": "prisma db seed",
+    "prisma:generate": "prisma generate"
   },
   "prisma": {
-    "seed": "node src/seed.js"
+    "seed": "ts-node prisma/seed.ts"
   },
   "author": "",
   "license": "ISC",
@@ -25,15 +26,20 @@
     "@babel/preset-env": "^7.20.2",
     "@babel/preset-typescript": "^7.18.6",
     "@faker-js/faker": "^7.6.0",
+    "@swc/cli": "^0.1.57",
+    "@swc/core": "^1.3.22",
     "@types/cors": "^2.8.13",
     "@types/jsonwebtoken": "^8.5.9",
     "@types/node": "^18.11.13",
     "babel-jest": "^29.3.1",
     "concurrently": "^7.6.0",
+    "cross-env": "^7.0.3",
     "jest-mock-extended": "^2.0.4",
     "nodemon": "^2.0.20",
     "prettier": "2.8.0",
     "ts-jest": "^29.0.3",
+    "ts-node": "^10.9.1",
+    "tsconfig-paths": "^4.1.1",
     "typescript": "^4.9.3"
   },
   "dependencies": {
