+++ src/exceptions/HttpException.ts
@@ -1,0 +1,17 @@
+export enum HttpCode {
+  NotFound = 404,
+  InternalServerError = 500,
+}
+
+export class HttpException {
+  public status: HttpCode
+  public message: string
+
+  constructor(
+    message: string,
+    status: HttpCode = HttpCode.InternalServerError
+  ) {
+    this.status = status
+    this.message = message
+  }
+}
