+++ src/middlewares/error.middleware.ts
@@ -1,5 +1,5 @@
 import { Request, Response, NextFunction } from 'express'
-import { HttpError } from '../types/errors'
+import { HttpException } from '@exceptions/HttpException'
 
 /**
  * Custom error handler to standardize error objects returned to
@@ -11,24 +11,25 @@ import { HttpError } from '../types/errors'
  * @param next NextFunction function provided by Express
  */
 function errorHandler(
-  err: TypeError | HttpError,
+  err: TypeError | HttpException,
   req: Request,
   res: Response,
   next: NextFunction
 ) {
   let customError = err
 
-  if (!(err instanceof HttpError)) {
-    customError = new HttpError(
+  if (!(err instanceof HttpException)) {
+    customError = new HttpException(
       'Oh no, this is embarrasing. We are having troubles my friend'
     )
   }
+  console.log(customError)
 
   // we are not using the next function to prevent from triggering
   // the default error-handler. However, make sure you are sending a
   // response to client to prevent memory leaks in case you decide to
   // NOT use, like in this example, the NextFunction .i.e., next(new Error())
-  res.status((customError as HttpError).status).send(customError)
+  res.status((customError as HttpException).status).send(customError)
 }
 
 export default errorHandler
